Remove helpers from redflags module
